Move _async_client.py to a _generated/ folder
