+++ src/huggingface_hub/__init__.py
@@ -203,12 +203,12 @@ _SUBMOD_ATTRS = {
         "ModelHubMixin",
         "PyTorchModelHubMixin",
     ],
-    "inference._async_client": [
-        "AsyncInferenceClient",
-    ],
     "inference._client": [
         "InferenceClient",
         "InferenceTimeoutError",
+    ],
+    "inference._generated._async_client": [
+        "AsyncInferenceClient",
     ],
     "inference_api": [
         "InferenceApi",
@@ -502,11 +502,11 @@ if TYPE_CHECKING:  # pragma: no cover
         ModelHubMixin,  # noqa: F401
         PyTorchModelHubMixin,  # noqa: F401
     )
-    from .inference._async_client import AsyncInferenceClient  # noqa: F401
     from .inference._client import (
         InferenceClient,  # noqa: F401
         InferenceTimeoutError,  # noqa: F401
     )
+    from .inference._generated._async_client import AsyncInferenceClient  # noqa: F401
     from .inference_api import InferenceApi  # noqa: F401
     from .keras_mixin import (
         KerasModelHubMixin,  # noqa: F401
