+++ utils/generate_async_inference_client.py
@@ -12,7 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Contains a tool to generate `src/huggingface_hub/inference/_async_client.py`."""
+"""Contains a tool to generate `src/huggingface_hub/inference/_generated/_async_client.py`."""
 import argparse
 import os
 import re
@@ -23,7 +23,9 @@ from typing import NoReturn
 from ruff.__main__ import find_ruff_bin
 
 
-ASYNC_CLIENT_FILE_PATH = Path(__file__).parents[1] / "src" / "huggingface_hub" / "inference" / "_async_client.py"
+ASYNC_CLIENT_FILE_PATH = (
+    Path(__file__).parents[1] / "src" / "huggingface_hub" / "inference" / "_generated" / "_async_client.py"
+)
 SYNC_CLIENT_FILE_PATH = Path(__file__).parents[1] / "src" / "huggingface_hub" / "inference" / "_client.py"
 
 
@@ -85,15 +87,16 @@ def check_async_client(update: bool) -> NoReturn:
 
             print(
                 "✅ AsyncInferenceClient source code has been updated in"
-                " `./src/huggingface_hub/inference/_async_client.py`.\n   Please make sure the changes are accurate"
-                " and commit them."
+                " `./src/huggingface_hub/inference/_generated/_async_client.py`.\n   Please make sure the changes are"
+                " accurate and commit them."
             )
             exit(0)
         else:
             print(
-                "❌ Expected content mismatch in `./src/huggingface_hub/inference/_async_client.py`.\n   It is most"
-                " likely that you modified some InferenceClient code and did not update the the AsyncInferenceClient"
-                " one.\n   Please run `make style` or `python utils/generate_async_inference_client.py --update`."
+                "❌ Expected content mismatch in `./src/huggingface_hub/inference/_generated/_async_client.py`.\n   It"
+                " is most likely that you modified some InferenceClient code and did not update the the"
+                " AsyncInferenceClient one.\n   Please run `make style` or `python"
+                " utils/generate_async_inference_client.py --update`."
             )
             exit(1)
 
@@ -103,8 +106,9 @@ def check_async_client(update: bool) -> NoReturn:
 
 def _add_warning_to_file_header(code: str) -> str:
     warning_message = (
-        "#\n# WARNING\n# This entire file has been generated automatically based on"
-        " `src/huggingface_hub/inference/_client.py`.\n# To re-generate it, run `make style` or `python"
+        "#\n# WARNING\n# This entire file has been adapted from the sync-client code in"
+        " `src/huggingface_hub/inference/_client.py`.\n# Any change in InferenceClient will be automatically reflected"
+        " in AsyncInferenceClient.\n# To re-generate the code, run `make style` or `python"
         " ./utils/generate_async_inference_client.py --update`.\n# WARNING"
     )
     return re.sub(
@@ -326,7 +330,10 @@ if __name__ == "__main__":
     parser.add_argument(
         "--update",
         action="store_true",
-        help="Whether to re-generate `./src/huggingface_hub/inference/_async_client.py` if a change is detected.",
+        help=(
+            "Whether to re-generate `./src/huggingface_hub/inference/_generated/_async_client.py` if a change is"
+            " detected."
+        ),
     )
     args = parser.parse_args()
